+++ Changes.md
@@ -5,6 +5,7 @@
 ### Bugs
 
 * Fixed bug in LocalCluster parsing critical_value property.
+* Fixed bug parsing tags in Source creation.
 
 ## 1.8.7 version